nginx-0.1.33-RELEASE import
    *) Bugfix: nginx could not be built with the --without-pcre parameter;
       the bug had appeared in 0.1.29.

    *) Bugfix: 3, 4, 7, and 8 the "proxy_set_header" directives in one
       level cause the bus fault on start up.

    *) Bugfix: the HTTP protocol was specified in the HTTPS redirects.

    *) Bugfix: if the "rewrite" directive used the captures inside the "if"
       directive, then the 500 error code was returned.


Former-commit-id: 4959ec482e14884e6c132ed0533c6d62bcc89405

+++ src/core/ngx_config.h
@@ -563,7 +563,7 @@ DECL|macro|NGX_MAX_UINT32_VALUE
 define|#
 directive|define
 name|NGX_MAX_UINT32_VALUE
-value|0xffffffffLL
+value|(uint32_t) 0xffffffffLL
 end_define
 
 begin_else
@@ -576,7 +576,7 @@ DECL|macro|NGX_MAX_UINT32_VALUE
 define|#
 directive|define
 name|NGX_MAX_UINT32_VALUE
-value|0xffffffff
+value|(uint32_t) 0xffffffff
 end_define
 
 begin_endif
